Change in gitQuizMain.c
Followed the comment on Line 1 which asked to add an else statement to show what will print if the user does not enter 'y'. The code that was added said that if the user does not enter 'y', it will print the following: You don't know how to use Git!